+++ src/Objects/Projectiles/ArrowProjectile.cs
@@ -3,7 +3,7 @@ namespace MyGame
 {
 	public class ArrowProjectile : Projectile
 	{
-		public ArrowProjectile (Direction direct, float x, float y, int damage) 
+		public ArrowProjectile (Direction direct, int x, int y, int damage) 
 			: base ("arrowProjectile", "spriteAnimation", direct, x, y, 5, damage)
 		{
 		}
